Remove PageIndexProvider
